Revert the move semantics and traits stuff (the 3 previous commits).
Although it's a shame to have to undo work, I felt that the move
semantics stuff was going in the wrong direction and adding too
much complexity to the compiler, so I decided to revert it.

Traits weren't a bad idea and I may come back to adding something
like traits in the future.